node: generate orchestrates through the service assembly seam
Move assemble_fresh into MiningService with a shared candidate_context
builder, and extend MempoolSnapshotSource with selection_snapshot so
the ordered-selection capture keeps its single pool read lock. The
node facade keeps generate's orchestration (solve, proposal
validation, authoritative submission) and its node-owned facts; the
dead coordinator mempool field and the duplicated MAX_BLOCK_*
constants leave with the logic.

Verified: cargo check -p bitcoin-rs-mining -p bitcoin-rs-node
-p bitcoin-rs-rpc; cargo test -p bitcoin-rs-node (556 lib + 49
mining integration + suites, 0 failed); cargo test -p bitcoin-rs-rpc
(11 suites, 0 failed).
